requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
billiard==4.2.1
black==24.10.0
boto3==1.38.35
botocore==1.38.35
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
loguru==0.7.3
Mako==1.3.10
mangum==0.19.0
MarkupSafe==3.0.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.0.1
prompt_toolkit==3.0.51
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.31.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
